feat: 검증된 GameResult 모델 기능 구현
