+++ README.md
@@ -67,14 +67,14 @@
 - Result
   - 결과는 게임 결과를 관리한다.
   - 결과는 참여자들을 가지고 있다.
-    - [ ] 참여자의 상태를 통해 게임 결과를 관리한다.
+    - [x] 참여자의 상태를 통해 게임 결과를 관리한다.
     - 참여자들의 블랙잭 결과에 따라 딜러와 플레이어의 승패를 결정한다.
-      - [ ] state를 비교해서 `버스트` 참여자를 `패배` 처리한다.
-      - [ ] state를 비고해서 딜러가 버스트 라면 `버스트 되지 않은` 참가자는 `승리` 처리한다. 
-      - [ ] 딜러의 점수를 기준으로 플레이어의 승패를 판단한다.
-        - [ ] 딜러보다 낮은 플레이어를 패배 처리한다.
-        - [ ] 딜러와 같은 플레이어를 무승부 처리한다.
-        - [ ] 딜러보다 높은 플레이어를 승리 처리한다.
+      - [x] state를 비교해서 `버스트` 참여자를 `패배` 처리한다.
+      - [x] state를 비고해서 딜러가 버스트 라면 `버스트 되지 않은` 참가자는 `승리` 처리한다. 
+      - [x] 딜러의 점수를 기준으로 플레이어의 승패를 판단한다.
+        - [x] 딜러보다 낮은 플레이어를 패배 처리한다.
+        - [x] 딜러와 같은 플레이어를 무승부 처리한다.
+        - [x] 딜러보다 높은 플레이어를 승리 처리한다.
 
 ### View
